# backend/src/main/resources/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Amena Backend API
  description: API specification for the Amena Backend Service
  version: 1.0.0
  contact:
    name: Amena Team
    email: support@amena.com
servers:
  - url: http://localhost:8081
    description: Development server
...
paths:
...
  /api/categories:
    get:
      tags:
        - Categories
      summary: List all categories
      operationId: getCategories
      responses:
        "200":
          description: List of all categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryResponse"
    post:
      tags:
        - Categories
      summary: Create a new category
      operationId: createCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryRequest"
      responses:
        "201":
          description: Category created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryResponse"
  /api/categories/{categoryId}:
    get:
      tags:
        - Categories
      summary: Get category by ID
      operationId: getCategoryById
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Category found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryResponse"
        "404":
          description: Category not found
    put:
      tags:
        - Categories
      summary: Update category
      operationId: updateCategory
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryRequest"
      responses:
        "200":
          description: Category updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryResponse"
        "404":
          description: Category not found
    delete:
      tags:
        - Categories
      summary: Delete category
      operationId: deleteCategory
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Category deleted successfully
        "404":
          description: Category not found
  /api/orders:
    get:
      tags:
        - Orders
      summary: List all orders with pagination and filtering
      operationId: getOrders
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: sortBy
          in: query
          schema:
            type: string
            default: "createdAt"
        - name: sortOrder
          in: query
          schema:
            type: string
            default: "desc"
      responses:
        "200":
          description: Paginated list of orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderPage"
    post:
      tags:
        - Orders
      summary: Create a new order
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
  /api/orders/{orderId}:
    get:
      tags:
        - Orders
      summary: Get order by ID
      operationId: getOrderById
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Order found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
        "404":
          description: Order not found
    delete:
      tags:
        - Orders
      summary: Delete order by ID
      operationId: deleteOrder
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Order deleted
        "404":
          description: Order not found
  /api/orders/{orderId}/status:
    put:
      tags:
        - Orders
      summary: Update order status
      operationId: updateOrderStatus
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdateRequest"
      responses:
        "200":
          description: Order status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
        "404":
          description: Order not found
  /api/coupons:
    get:
      tags:
        - Coupons
      summary: List all coupons
      operationId: getCoupons
      responses:
        "200":
          description: List of all coupons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CouponResponse"
    post:
      tags:
        - Coupons
      summary: Create a new coupon
      operationId: createCoupon
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponRequest"
      responses:
        "201":
          description: Coupon created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CouponResponse"
        "409":
          description: Coupon code already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/coupons/{couponId}:
    get:
      tags:
        - Coupons
      summary: Get coupon by ID
      operationId: getCouponById
      parameters:
        - name: couponId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Coupon found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CouponResponse"
        "404":
          description: Coupon not found
    put:
      tags:
        - Coupons
      summary: Update coupon
      operationId: updateCoupon
      parameters:
        - name: couponId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponRequest"
      responses:
        "200":
          description: Coupon updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CouponResponse"
        "404":
          description: Coupon not found
    delete:
      tags:
        - Coupons
      summary: Delete coupon
      operationId: deleteCoupon
      parameters:
        - name: couponId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Coupon deleted successfully
        "404":
          description: Coupon not found
  /api/offres:
    get:
      tags:
        - Offres
      summary: List all offres
      operationId: getOffres
      responses:
        "200":
          description: List of all offres
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OffreResponse"
    post:
      tags:
        - Offres
      summary: Create a new offre
      operationId: createOffre
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OffreRequest"
      responses:
        "201":
          description: Offre created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OffreResponse"
  /api/offres/{offreId}:
    get:
      tags:
        - Offres
      summary: Get offre by ID
      operationId: getOffreById
      parameters:
        - name: offreId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Offre found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OffreResponse"
        "404":
          description: Offre not found
    put:
      tags:
        - Offres
      summary: Update offre
      operationId: updateOffre
      parameters:
        - name: offreId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OffreRequest"
      responses:
        "200":
          description: Offre updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OffreResponse"
        "404":
          description: Offre not found
    delete:
      tags:
        - Offres
      summary: Delete offre
      operationId: deleteOffre
      parameters:
        - name: offreId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Offre deleted successfully
        "404":
          description: Offre not found
  /api/bannieres:
    get:
      tags:
        - Bannieres
      summary: List all bannieres
      operationId: getBannieres
      responses:
        "200":
          description: List of all bannieres
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BanniereResponse"
    post:
      tags:
        - Bannieres
      summary: Create a new banniere
      operationId: createBanniere
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BanniereRequest"
      responses:
        "201":
          description: Banniere created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BanniereResponse"
  /api/bannieres/upload-image:
    post:
      tags:
        - Bannieres
      summary: Upload an image for a banniere
      operationId: uploadBanniereImage
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: Image uploaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  imageUrl:
                    type: string
  /api/bannieres/{banniereId}:
    get:
      tags:
        - Bannieres
      summary: Get banniere by ID
      operationId: getBanniereById
      parameters:
        - name: banniereId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Banniere found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BanniereResponse"
        "404":
          description: Banniere not found
    put:
      tags:
        - Bannieres
      summary: Update banniere
      operationId: updateBanniere
      parameters:
        - name: banniereId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BanniereRequest"
      responses:
        "200":
          description: Banniere updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BanniereResponse"
        "404":
          description: Banniere not found
    delete:
      tags:
        - Bannieres
      summary: Delete banniere
      operationId: deleteBanniere
      parameters:
        - name: banniereId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Banniere deleted successfully
        "404":
          description: Banniere not found
...
components:
  schemas:
...
    BanniereRequest:
      type: object
      required:
        - title
        - status
      properties:
        title:
          type: string
        imageUrl:
          type: string
        targetUrl:
          type: string
        position:
          type: integer
        status:
          type: string
          enum: [Actif, Inactif]
    BanniereResponse:
      type: object
      required:
        - id
        - title
        - status
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        imageUrl:
          type: string
        targetUrl:
          type: string
        position:
          type: integer
        status:
          type: string
          enum: [Actif, Inactif]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    OrderPage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/OrderResponse"
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean
...
    CategoryResponse:
      type: object
      required:
        - id
        - name
        - slug
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        parentId:
          type: integer
          format: int64
        subcategory:
          type: string
        imageUrl:
          type: string
        isActive:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
        productCount:
          type: integer
    OrderItemResponse:
      type: object
      required:
        - productName
        - quantity
        - unitPrice
        - totalPrice
      properties:
        productName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
          format: double
        totalPrice:
          type: number
          format: double
    OrderResponse:
      type: object
      required:
        - id
        - orderNumber
        - clientName
        - totalAmount
        - status
        - createdAt
        - items
      properties:
        id:
          type: integer
          format: int64
        orderNumber:
          type: string
        clientName:
          type: string
        clientPhone:
          type: string
        userId:
          type: integer
          format: int64
        totalAmount:
          type: number
          format: double
        shippingAmount:
          type: number
          format: double
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        address:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemResponse"
    OrderStatusUpdateRequest:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          description: pending, processing, shipped, delivered, cancelled, refunded
    CreateOrderRequest:
      type: object
      required:
        - subtotal
        - totalAmount
        - items
      properties:
        clientInfo:
          type: string
        subtotal:
          type: number
          format: double
        shippingAmount:
          type: number
          format: double
        discountAmount:
          type: number
          format: double
        totalAmount:
          type: number
          format: double
        couponCode:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/CreateOrderItemRequest"
    CreateOrderItemRequest:
      type: object
      required:
        - productId
        - productName
        - quantity
        - unitPrice
      properties:
        productId:
          type: integer
          format: int64
        productName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
          format: double
    CategoryRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        parentId:
          type: integer
          format: int64
        subcategory:
          type: string
        imageUrl:
          type: string
        isActive:
          type: boolean
        sortOrder:
          type: integer
...
    ErrorResponse:
      type: object
      required:
        - timestamp
        - status
        - error
        - message
        - path
      properties:
        timestamp:
          type: string
          format: date-time
          description: Error occurrence timestamp
        status:
          type: integer
          description: HTTP status code
        error:
          type: string
          description: Error type
        message:
          type: string
          description: Detailed error message
        path:
          type: string
          description: Request path that caused the error
    CouponResponse:
      type: object
      required:
        - id
        - code
        - discountType
        - discountValue
        - isActive
      properties:
        id:
          type: integer
          format: int64
        code:
          type: string
        description:
          type: string
        discountType:
          type: string
          description: "percentage or fixed_amount"
        discountValue:
          type: number
          format: double
        minimumOrderAmount:
          type: number
          format: double
        maximumDiscount:
          type: number
          format: double
        usageLimit:
          type: integer
        usedCount:
          type: integer
        startsAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    CouponRequest:
      type: object
      required:
        - code
        - discountType
        - discountValue
      properties:
        code:
          type: string
        description:
          type: string
        discountType:
          type: string
          description: "percentage or fixed_amount"
        discountValue:
          type: number
          format: double
        minimumOrderAmount:
          type: number
          format: double
        maximumDiscount:
          type: number
          format: double
        usageLimit:
          type: integer
        startsAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        isActive:
          type: boolean
    OffreResponse:
      type: object
      required:
        - id
        - label
        - discountPercentage
        - isActive
      properties:
        id:
          type: integer
          format: int64
        label:
          type: string
        categoryId:
          type: integer
          format: int64
        categoryName:
          type: string
          description: Resolved category name for display
        discountPercentage:
          type: number
          format: double
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    OffreRequest:
      type: object
      required:
        - label
        - discountPercentage
      properties:
        label:
          type: string
        categoryId:
          type: integer
          format: int64
        discountPercentage:
          type: number
          format: double
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        isActive:
          type: boolean
...